Get DESeq2 Results

This function obtains a set of comparisons from a DESeq2 analysis, given a named list of contrasts. It allows additional model parameters to be specified and a design matrix to be manually adjusted.

Usage

get_DESeq2_res(
  dds,
  contrasts,
  res.list = list(),
  user.mat = FALSE,
  block = NULL,
  design = NULL,
  alpha = 0.05,
  lfc.th = c(log2(1.25), log2(1.5)),
  shrink.method = "ashr",
  norm.ercc = FALSE,
  add.rowData = NULL,
  BPPARAM = NULL
)

Arguments

dds

An object of class DESeqDataSet.

contrasts

A named list of contrasts.

res.list

A named list to hold DESeq2 result tables. Allows the function to be run multiple times if needed and append to the same list. Defaults to an empty list.

user.mat

A logical indicating whether a user-specified model matrix is provided. Defaults to FALSE.

block

A vector of additional terms to be considered in the model, beyond the main effect. Defaults to NULL.

design

The design formula or matrix. If a matrix is provided, ensure 'user.mat' is set to TRUE. Defaults to NULL.

alpha

The significance level for hypothesis testing. Defaults to 0.05.

lfc.th

A numeric vector of log2 fold-change thresholds. Defaults to `c(log2(1.25), log2(1.5))“.

shrink.method

The method used for shrinkage estimation. Defaults to "ashr".

norm.ercc

A logical indicating whether to normalize to ERCC spike-ins.

add.rowData

A vector of column names from the rowData slot of the DESeqDataSet to be added to the results table. Defaults to NULL.

BPPARAM

The BiocParallelParam object specifying the parallel back-end to be used. Defaults to NULL.

Value

A named list of DESeq2::DESeqResults objects for the specified contrasts. If add.rowData is supplied, these will be returned as S4Vectors::DFrame objects instead.

Author

Jared Andrews

Examples

library(DESeq2)
#> Loading required package: S4Vectors
#> Loading required package: stats4
#> Loading required package: BiocGenerics
#> 
#> Attaching package: ‘BiocGenerics’
#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs
#> The following objects are masked from ‘package:base’:
#> 
#>     Filter, Find, Map, Position, Reduce, anyDuplicated, aperm, append,
#>     as.data.frame, basename, cbind, colnames, dirname, do.call,
#>     duplicated, eval, evalq, get, grep, grepl, intersect, is.unsorted,
#>     lapply, mapply, match, mget, order, paste, pmax, pmax.int, pmin,
#>     pmin.int, rank, rbind, rownames, sapply, saveRDS, setdiff, table,
#>     tapply, union, unique, unsplit, which.max, which.min
#> 
#> Attaching package: ‘S4Vectors’
#> The following object is masked from ‘package:utils’:
#> 
#>     findMatches
#> The following objects are masked from ‘package:base’:
#> 
#>     I, expand.grid, unname
#> Loading required package: IRanges
#> Loading required package: GenomicRanges
#> Loading required package: GenomeInfoDb
#> Loading required package: SummarizedExperiment
#> Loading required package: MatrixGenerics
#> Loading required package: matrixStats
#> 
#> Attaching package: ‘MatrixGenerics’
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     colAlls, colAnyNAs, colAnys, colAvgsPerRowSet, colCollapse,
#>     colCounts, colCummaxs, colCummins, colCumprods, colCumsums,
#>     colDiffs, colIQRDiffs, colIQRs, colLogSumExps, colMadDiffs,
#>     colMads, colMaxs, colMeans2, colMedians, colMins, colOrderStats,
#>     colProds, colQuantiles, colRanges, colRanks, colSdDiffs, colSds,
#>     colSums2, colTabulates, colVarDiffs, colVars, colWeightedMads,
#>     colWeightedMeans, colWeightedMedians, colWeightedSds,
#>     colWeightedVars, rowAlls, rowAnyNAs, rowAnys, rowAvgsPerColSet,
#>     rowCollapse, rowCounts, rowCummaxs, rowCummins, rowCumprods,
#>     rowCumsums, rowDiffs, rowIQRDiffs, rowIQRs, rowLogSumExps,
#>     rowMadDiffs, rowMads, rowMaxs, rowMeans2, rowMedians, rowMins,
#>     rowOrderStats, rowProds, rowQuantiles, rowRanges, rowRanks,
#>     rowSdDiffs, rowSds, rowSums2, rowTabulates, rowVarDiffs, rowVars,
#>     rowWeightedMads, rowWeightedMeans, rowWeightedMedians,
#>     rowWeightedSds, rowWeightedVars
#> Loading required package: Biobase
#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.
#> 
#> Attaching package: ‘Biobase’
#> The following object is masked from ‘package:MatrixGenerics’:
#> 
#>     rowMedians
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     anyMissing, rowMedians
dds_de <- makeExampleDESeqDataSet(n = 100, m = 12, betaSD = 2) # DE genes
rownames(dds_de) <- paste0("gene", 1001:1100)
dds <- makeExampleDESeqDataSet(n = 1000, m = 12) # Non-DE genes
dds <- rbind(dds, dds_de)
dds <- DESeq(dds)
#> estimating size factors
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#> final dispersion estimates
#> fitting model and testing
contrasts <- list("condition" = c("condition", "A", "B"))
res <- get_DESeq2_res(dds, contrasts)
#> Loading required package: ashr
#> 
#> Design for condition_A_vs_B is ~condition
#> using pre-existing size factors
#> estimating dispersions
#> found already estimated dispersions, replacing these
#> gene-wise dispersion estimates
#> found already estimated gene-wise dispersions, removing these
#> mean-dispersion relationship
#> final dispersion estimates
#> found already estimated dispersions, removing these
#> fitting model and testing
#> using 'ashr' for LFC shrinkage. If used in published research, please cite:
#>     Stephens, M. (2016) False discovery rates: a new deal. Biostatistics, 18:2.
#>     https://doi.org/10.1093/biostatistics/kxw041
#> using 'ashr' for LFC shrinkage. If used in published research, please cite:
#>     Stephens, M. (2016) False discovery rates: a new deal. Biostatistics, 18:2.
#>     https://doi.org/10.1093/biostatistics/kxw041
#> computing FSOS 'false sign or small' s-values (T=0.322)
#> using 'ashr' for LFC shrinkage. If used in published research, please cite:
#>     Stephens, M. (2016) False discovery rates: a new deal. Biostatistics, 18:2.
#>     https://doi.org/10.1093/biostatistics/kxw041
#> computing FSOS 'false sign or small' s-values (T=0.585)

names(res)
#> [1] "condition"                        "condition-shLFC"                 
#> [3] "condition-shLFC0.321928094887362" "condition-LFC0.321928094887362"  
#> [5] "condition-shLFC0.584962500721156" "condition-LFC0.584962500721156"  
head(res[[1]])
#> log2 fold change (MLE): condition A vs B 
#> Wald test p-value: condition A vs B 
#> DataFrame with 6 rows and 6 columns
#>        baseMean log2FoldChange     lfcSE      stat    pvalue      padj
#>       <numeric>      <numeric> <numeric> <numeric> <numeric> <numeric>
#> gene1   4.12848     -0.2906540  0.969207 -0.299888  0.764262  0.984393
#> gene2  17.06387     -0.2622629  0.427627 -0.613298  0.539679  0.972929
#> gene3 178.18356      0.1426357  0.334604  0.426282  0.669902  0.974509
#> gene4   5.60499      0.4369778  0.766539  0.570066  0.568633  0.974509
#> gene5  90.75194      0.1498189  0.283374  0.528697  0.597016  0.974509
#> gene6   5.55006     -0.0913117  0.773288 -0.118082  0.906002  0.992109